src/Web-App/requirements.txt
python 3.10.10
aws-cli/1.27.74
botocore/1.29.74
boto3-1.26.74
Flask 2.2.3
Werkzeug 2.2.3
